Move "create wallet" command into "wallet"
