+++ neo/Prompt/Commands/Wallet.py
@@ -7,16 +7,20 @@ from neo.Core.TX.TransactionAttribute import TransactionAttribute, TransactionAt
 from neo.SmartContract.ContractParameterContext import ContractParametersContext
 from neo.Network.NodeLeader import NodeLeader
 from neo.Prompt.Utils import get_asset_id, get_from_addr, get_arg
+from neo.Wallets.utils import to_aes_key
+from neo.Implementations.Wallets.peewee.UserWallet import UserWallet
 from neocore.Fixed8 import Fixed8
 from neocore.UInt160 import UInt160
 from prompt_toolkit import prompt
 import binascii
 import json
+import os
 import math
 from neo.Implementations.Wallets.peewee.Models import Account
 from neo.Prompt.CommandBase import CommandBase, CommandDesc, ParameterDesc
 from neo.Prompt.PromptData import PromptData
 from neo.logging import log_manager
+
 
 logger = log_manager.getLogger()
 
@@ -25,6 +29,7 @@ class CommandWallet(CommandBase):
     def __init__(self):
         super().__init__()
 
+        self.register_sub_command('create', CommandWalletCreate())
         self.register_sub_command(['v', '--v', 'verbose'], CommandWalletVerbose())
         self.register_sub_command('migrate', CommandWalletMigrate())
         self.register_sub_command('create_addr', CommandWalletCreateAddress())
@@ -34,11 +39,16 @@ class CommandWallet(CommandBase):
 
     def execute(self, arguments):
         wallet = PromptData.Wallet
+        item = get_arg(arguments)
+
+        # Create and Open must be handled specially.
+        if item == 'create':
+            self.execute_sub_command(item, arguments[1:])
+            return
+
         if not wallet:
             print("Please open a wallet")
             return
-
-        item = get_arg(arguments)
 
         if not item:
             print("Wallet %s " % json.dumps(wallet.ToJson(), indent=4))
@@ -48,6 +58,57 @@ class CommandWallet(CommandBase):
             self.execute_sub_command(item, arguments[1:])
         except KeyError:
             print(f"Wallet: {item} is an invalid parameter")
+
+
+class CommandWalletCreate(CommandBase):
+
+    def __init__(self):
+        super().__init__()
+
+    @classmethod
+    def execute(cls, arguments):
+        path = get_arg(arguments, 0)
+
+        if path:
+            if os.path.exists(path):
+                print("File already exists")
+                return
+
+            passwd1 = prompt("[password]> ", is_password=True)
+            passwd2 = prompt("[password again]> ", is_password=True)
+
+            if passwd1 != passwd2 or len(passwd1) < 10:
+                print("Please provide matching passwords that are at least 10 characters long")
+                return
+
+            password_key = to_aes_key(passwd1)
+
+            try:
+                PromptData.Wallet = UserWallet.Create(path=path, password=password_key)
+                contract = PromptData.Wallet.GetDefaultContract()
+                key = PromptData.Wallet.GetKey(contract.PublicKeyHash)
+                print("Wallet %s" % json.dumps(PromptData.Wallet.ToJson(), indent=4))
+                print("Pubkey %s" % key.PublicKey.encode_point(True))
+            except Exception as e:
+                print("Exception creating wallet: %s" % e)
+                PromptData.Wallet = None
+                if os.path.isfile(path):
+                    try:
+                        os.remove(path)
+                    except Exception as e:
+                        print("Could not remove {}: {}".format(path, e))
+                return
+
+            if PromptData.Wallet:
+                PromptData.Prompt.start_wallet_loop()
+
+        else:
+            print("Please specify a path")
+
+    @classmethod
+    def command_desc(self):
+        p1 = ParameterDesc('path', 'path to store the wallet file')
+        return CommandDesc('create', 'creates a new NEO wallet address', [p1])
 
 
 class CommandWalletVerbose(CommandBase):
